ImageEditor updated, AvatarEditor added
